+++ crc32_rohd/lib/interfaces.dart
